as: Move PubSub ID functions to main package
