+++ pkg/applicationserver/io/pubsub/unique.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package unique
+package pubsub
 
 import (
 	"fmt"
@@ -21,13 +21,13 @@ import (
 
 var uniqueIDPattern = regexp.MustCompile("(.*)\\:(.*)")
 
-// ID generates an unique ID based on an application unique ID and a PubSub ID.
-func ID(appUID, id string) string {
+// PubSubUID generates an unique ID based on an application unique ID and a PubSub ID.
+func PubSubUID(appUID, id string) string {
 	return fmt.Sprintf("%s:%s", appUID, id)
 }
 
-// ToUIDs parses a unique ID generated by `ID` and returns the application unique ID and the PubSub ID.
-func ToUIDs(uid string) (string, string) {
+// SplitPubSubUID parses a unique ID generated by `ID` and returns the application unique ID and the PubSub ID.
+func SplitPubSubUID(uid string) (string, string) {
 	matches := uniqueIDPattern.FindStringSubmatch(uid)
 	if len(matches) != 3 {
 		panic(fmt.Sprintf("invalid uniqueID `%s` with matches %v", uid, matches))
